Debugger service pack - fixed bp in static blocks and breakpoints at early code execution are working now (#186)
* Debugger: fixed not working breakpoints in static blocks

* debugger: it suspends every thread now once JDPW client/server is connected and sends VM_START event. As it there were breakpoint skipped due fact that set breakpoint request came late after code was already executed

+++ plugins/debugger/src/main/java/org/robovm/debugger/jdwp/JdwpConsts.java
@@ -124,9 +124,9 @@ public class JdwpConsts {
     }
 
     public static class SuspendPolicy {
-        public static final int NONE = 0;
-        public static final int EVENT_THREAD = 1;
-        public static final int ALL = 2;
+        public static final byte NONE = 0;
+        public static final byte EVENT_THREAD = 1;
+        public static final byte ALL = 2;
     }
 
     public static class TypeTag {
